Fixed hand border color always being black
Added handBorderColor setting if block from init to the function where it's rendered. Sort of dumb workaround but I don't know why it wasn't working properly. Maybe something to do with memory? Also increased the buffer for digital time layer from 8 to 9, because that was an issue for some reason. I don't know. I need to refactor all of this code sometime because it's a god damn mess.